refactor elements
